Separated Interactables and GUI
interactable and info GUI scripts are their own module
