+++ Assets/Scripts/AddContainerEvent.cs
@@ -7,7 +7,9 @@ using UnityEngine.Events;
  * Created by Ethan Wolfe
  * Used for pickup and place event triggers
  */
+namespace BaristaSimulator
+{
+    public class AddContainerEvent : UnityEvent<LiquidContainer> { }
 
-public class AddContainerEvent : UnityEvent<LiquidContainer> { }
-
-public class RemoveContainerEvent : UnityEvent { };
+    public class RemoveContainerEvent : UnityEvent { };
+}
