+++ Assets/Scripts/Interactable Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f8bb17bcd8df8c641b4b810fd3ee6fe0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
