+++ Assets/Scripts/Interactable Objects/Espresso Machine.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc685e70494bd7145831f41869ca648f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
