+++ Assets/Scripts/Interactable Objects/Espresso Machine/CoffeeCup.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace BaristaSimulator
+{
+    public class CoffeeCup : LiquidContainer
+    {
+
+    }
+}
