+++ Assets/Scripts/Interactable Objects/Espresso Machine/CoffeeCup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9b07f40f50c32a144ba365a3c53df2a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
