+++ Assets/Scripts/Interactable Objects/Espresso Machine/DripTray.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 281c7943c37280547b976dfa415f78d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
