+++ Assets/Scripts/Interactable Objects/Espresso Machine/PortableFilter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0ea719babe493a144919ab7cf95769d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
