+++ Assets/Scripts/Interactable Objects/Espresso Machine/WaterContainer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6209d0bd17c4d6444956cf13503262cb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
