+++ Assets/Scripts/Interactable Objects/Faucet.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8843c26ff5a15534c9daeffa638f8010
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
