+++ Assets/Scripts/Interactable Objects/InformationInterface.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 77f9451d138776941a0123de876532f8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
