+++ Assets/Scripts/LiquidContainerReference.cs
@@ -5,10 +5,15 @@ using UnityEngine;
  * Created by Ethan Wolfe
  * A simple script so that a liquid collider plane can have an easy link to the liquid container it resides in
  */
-public class LiquidContainerReference : MonoBehaviour {
-    [SerializeField] private LiquidContainer _container;
+namespace BaristaSimulator
+{
+    public class LiquidContainerReference : MonoBehaviour
+    {
+        [SerializeField] private LiquidContainer _container;
 
-    public LiquidContainer GetLiquidContianer() {
-        return _container;
+        public LiquidContainer GetLiquidContianer()
+        {
+            return _container;
+        }
     }
 }
