+++ Assets/Scripts/Player/IHoverableEnter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ded02e0f9ea374a4aaead23e2319d999
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
