+++ Assets/Scripts/Player/IHoverableExit.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace BaristaSimulator
+{
+    public interface IHoverableExit
+    {
+        void OnHoverExit();
+    }
+}
