+++ Assets/Scripts/Player/IHoverableExit.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fef5703bbac773f43927584aafb6154f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
