+++ Assets/Scripts/Player/IInteractable.cs
@@ -1,0 +1,9 @@
+
+
+namespace BaristaSimulator
+{
+    public interface IInteractable 
+    {
+        void OnInteract();
+    }
+}
