+++ Assets/Scripts/SpawnOnInteract.cs
@@ -4,33 +4,41 @@ using UnityEngine.Events;
  * Created by Ethan Wolfe
  * Will create a new instance of an item and also create a new instance of the material so that they don't interfere
  */
-public class SpawnOnInteract : IInteractable {
-    [SerializeField] private HandManager hand;
-    [SerializeField] private GameObject prefab;
+namespace BaristaSimulator
+{
+    public class SpawnOnInteract : MonoBehaviour, IInteractable, IHoverableEnter
+    {
+        [SerializeField] private HandManager hand;
+        [SerializeField] private GameObject prefab;
 
-    [SerializeField] protected UnityEvent<HeldObject> pickUpEvent; 
+        [SerializeField] protected UnityEvent<HeldObject> pickUpEvent;
 
-    public override void OnStartHover() {
-        base.OnStartHover();
-        if (!hand.HandEmpty()) {
-            interactionGUI.HideInfo();
+        [SerializeField] private InformationInterface _info;
+
+        public void OnHoverEnter()
+        {
+            if (!hand.HandEmpty())
+            {
+                _info.interactionGUI.HideInfo();
+            }
         }
-    }
 
-    public override bool OnInteract() {
-        if (hand.HandEmpty()) {
-            GameObject newObject = Instantiate(prefab);
-            HeldObject heldObject = newObject.GetComponent<HeldObject>();
-            
-            LiquidContainer container = newObject.GetComponent<LiquidContainer>();
-            GameObject liquidObject = container.GetLiquidObject();
-            Renderer renderer = liquidObject.GetComponent<Renderer>();
-            Material materialInstance = Instantiate(renderer.material);
-            container.UpdateMaterial(renderer, materialInstance);
+        public void OnInteract()
+        {
+            if (hand.HandEmpty())
+            {
+                GameObject newObject = Instantiate(prefab);
+                HeldObject heldObject = newObject.GetComponent<HeldObject>();
 
-            hand.PickUpItem(heldObject);
-            pickUpEvent.Invoke(heldObject);
+                LiquidContainer container = newObject.GetComponent<LiquidContainer>();
+                GameObject liquidObject = container.GetLiquidObject();
+                Renderer renderer = liquidObject.GetComponent<Renderer>();
+                Material materialInstance = Instantiate(renderer.material);
+                container.UpdateMaterial(renderer, materialInstance);
+
+                hand.PickUpItem(heldObject);
+                pickUpEvent.Invoke(heldObject);
+            }
         }
-        return false;
     }
 }
